requirements.txt
discord.py[voice]>=2.3.2
python-dotenv>=1.0.0
yt-dlp
PyNaCl>=1.5.0